Improves and finishes the ATM problem.
* Added tests.
    * Added a TestBase class containing methods for testing
      main methods, as well as an overridden assertEquals.
* Improves error checking and reporting.

+++ codeChef/atm/Account.java
@@ -21,7 +21,7 @@ public class Account {
         return owner;
     }
 
-    public Withdraw withdraw(double amount) {
+    public Withdraw withdraw(int amount) {
         return new Withdraw(balance, amount);
     }
